update mobile sidebar nav
