+++ src/components/Navigation/navigation.module.css
@@ -99,18 +99,28 @@
     position: fixed;
     top: 0;
     right: 0;
-    width: min(300px, 100dvw);
+    /* The sheet bleeds into the right inset, so its own padding keeps content clear of it */
+    width: min(calc(300px + env(safe-area-inset-right)), 100dvw);
     height: 100dvh;
-    padding: 30px 28px;
+    padding: calc(30px + env(safe-area-inset-top)) calc(28px + env(safe-area-inset-right))
+      calc(30px + env(safe-area-inset-bottom)) 28px;
     display: flex;
     flex-direction: column;
-    gap: 28px;
+    gap: 24px;
     transform: translateX(100%);
     transition:
       transform 0.3s ease,
       box-shadow 0.3s ease;
     z-index: 95;
+    /* Scrolls on its own when the links plus actions outgrow the viewport, without chaining to the page */
     overflow-y: auto;
+    overscroll-behavior: contain;
+    -webkit-overflow-scrolling: touch;
+
+    /* Sections must keep their height so the sheet scrolls instead of squeezing the buttons */
+    & > * {
+      flex: 0 0 auto;
+    }
   }
 
   .mobileMenuOpen {
@@ -130,20 +140,25 @@
   }
 
   .mobileActions {
+    border-top: 1px solid var(--border-glass);
     display: flex;
     flex-direction: column;
-    gap: 16px;
-    margin-top: auto;
-  }
+    gap: 12px;
+    padding-top: 16px;
 
-  .mobileActions .actionIconLink {
-    width: 44px;
-    height: 44px;
-  }
+    .mobileSocials {
+      display: flex;
+      gap: 12px;
 
-  .mobileActions .loginButton {
-    width: 100%;
-    text-align: center;
+      .socialButton {
+        flex: 1;
+      }
+    }
+
+    .loginButton {
+      width: 100%;
+      text-align: center;
+    }
   }
 
   .mobileBackdrop {
