+++ src/components/button/button.tsx
@@ -19,7 +19,8 @@ export type ButtonProps = {
   id?: string;
   /** Loading state; Shows spinner inside the button and disables it */
   loading?: boolean;
-  onClick?: MouseEventHandler<HTMLButtonElement>;
+  /** Fires on both render paths: the `<button>`, and the `<Link>` when `href` is set. */
+  onClick?: MouseEventHandler<HTMLButtonElement | HTMLAnchorElement>;
   target?: '_blank' | '_self';
   size?: 'link' | 'xs' | 'sm' | 'sm-const' | 'md' | 'md-const' | 'lg' | 'lg-const';
   type?: 'button' | 'submit' | 'reset';
@@ -71,7 +72,7 @@ const Button = forwardRef<HTMLButtonElement, ButtonProps>(
       />
     ) : null;
 
-    const handleClick = async (event: React.MouseEvent<HTMLButtonElement>) => {
+    const handleClick = async (event: React.MouseEvent<HTMLButtonElement | HTMLAnchorElement>) => {
       if (onClick) {
         if (autoLoading) {
           setInnerLoading(true);
@@ -85,7 +86,13 @@ const Button = forwardRef<HTMLButtonElement, ButtonProps>(
 
     if (href) {
       return (
-        <Link className={classes} href={isDisabled ? '' : href} id={id} target={target}>
+        <Link
+          className={classes}
+          href={isDisabled ? '' : href}
+          id={id}
+          onClick={onClick}
+          target={target}
+        >
           {spinner}
           {contentBefore}
           {children}
